Fix for PR (description)
- Microgame Folder correct name
- Every non-script asset got DrinkContest+X name
- Script rename DrinkContest
- Folder empty are remove
---------------------
- The game is now 8 beats
- The difficulty is hard but possible in speed 6 stage 3.
- All variable are now under the form WordLikeThis
- The GetComponent<> is use only when I need to change the sprite (3 time by game).
------------------------------
+ The game still work !
