+++ Assets/Microgames/DrinkContest/Scenes/DrinkContest2.unity
@@ -310,7 +310,7 @@ GameObject:
   m_Component:
   - component: {fileID: 216393195}
   - component: {fileID: 216393194}
-  - component: {fileID: 216393196}
+  - component: {fileID: 216393193}
   m_Layer: 0
   m_Name: Cara_1
   m_TagString: Untagged
@@ -318,6 +318,31 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
+--- !u!114 &216393193
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 216393192}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 822901568e446cf4ba4fe7b3441596a2, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  ObjectLiquide1: {fileID: 62205794}
+  ObjectLiquide2: {fileID: 88619309}
+  ObjectCara1: {fileID: 216393192}
+  ObjectCara2: {fileID: 600339549}
+  Cara1Anim1: {fileID: 21300000, guid: 000dbc023d863df43889d0675e35d663, type: 3}
+  Cara1Anim2: {fileID: 21300002, guid: 000dbc023d863df43889d0675e35d663, type: 3}
+  Cara1Anim3: {fileID: 21300004, guid: 000dbc023d863df43889d0675e35d663, type: 3}
+  Cara2Anim1: {fileID: 21300000, guid: bc0867dddb538094dad6fde5cef1cadb, type: 3}
+  Cara2Anim2: {fileID: 21300002, guid: bc0867dddb538094dad6fde5cef1cadb, type: 3}
+  Cara2Anim3: {fileID: 21300004, guid: bc0867dddb538094dad6fde5cef1cadb, type: 3}
+  percentbottle1: 100
+  percentbottle2: 100
+  PercentLooseClick: 10
+  DrinkSound: {fileID: 8300000, guid: aec6381c3115d314b82b59a34076c4cb, type: 3}
 --- !u!212 &216393194
 SpriteRenderer:
   m_ObjectHideFlags: 0
@@ -375,31 +400,6 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &216393196
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 216393192}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 26557711dfa72b34a93fe1306a9d3255, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  ObjectLiquide1: {fileID: 62205794}
-  ObjectLiquide2: {fileID: 88619309}
-  ObjectCara1: {fileID: 216393192}
-  ObjectCara2: {fileID: 600339549}
-  Cara1_Anim1: {fileID: 21300000, guid: 000dbc023d863df43889d0675e35d663, type: 3}
-  Cara1_Anim2: {fileID: 21300002, guid: 000dbc023d863df43889d0675e35d663, type: 3}
-  Cara1_Anim3: {fileID: 21300004, guid: 000dbc023d863df43889d0675e35d663, type: 3}
-  Cara2_Anim1: {fileID: 21300000, guid: bc0867dddb538094dad6fde5cef1cadb, type: 3}
-  Cara2_Anim2: {fileID: 21300002, guid: bc0867dddb538094dad6fde5cef1cadb, type: 3}
-  Cara2_Anim3: {fileID: 21300004, guid: bc0867dddb538094dad6fde5cef1cadb, type: 3}
-  percent_bottle1: 100
-  percent_bottle2: 100
-  percent_loose_click: 2.5
-  DrinkSound: {fileID: 8300000, guid: aec6381c3115d314b82b59a34076c4cb, type: 3}
 --- !u!1 &292370318
 GameObject:
   m_ObjectHideFlags: 0
