+++ Assets/Microgames/DrinkContest/Scenes/DrinkContest2.unity.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 75c2430721a57ca44b1bc637b53c585b
+guid: 62110e6912e05f0439408351bd832272
 timeCreated: 1476326322
 licenseType: Pro
 DefaultImporter:
