+++ Assets/Microgames/DrinkContest/Scenes/DrinkContest3.unity.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3e6318cc4a7674f41aa29a39e22c520e
+guid: a54da52aeb0e07d4b88dc6d539e9cf87
 timeCreated: 1476326322
 licenseType: Pro
 DefaultImporter:
