+++ Assets/Microgames/DrinkContest/Scripts/DrinkContestScript.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 822901568e446cf4ba4fe7b3441596a2
+timeCreated: 1545060494
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
