+++ Assets/Microgames/DrinkContest/Sprites.meta
