+++ Assets/Microgames/DrinkContest/Sprites/DrinkContestCara1.png.meta
