+++ Assets/Microgames/DrinkContest/Sprites/DrinkContestCara2.png
