+++ Assets/Microgames/DrinkContest/Sprites/DrinkContestDecors.png.meta
