+++ Assets/Microgames/DrinkContest/Sprites/DrinkContestLiquide.png.meta
