+++ Assets/Microgames/DrinkContest/Traits.asset
@@ -14,7 +14,7 @@ MonoBehaviour:
   _controlScheme: 0
   _hideCursor: 1
   _cursorLockState: 0
-  _duration: 1
+  _duration: 0
   _command: Drink fast (Space bar) !
   _defaultVictory: 0
   _victoryVoiceDelay: 0
