--- Assets/Microgames/DrinkingParty/Animation.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: b9c2a15f31dd06f4aa31af08055a162c
-folderAsset: yes
-timeCreated: 1536608453
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
