--- Assets/Microgames/DrinkingParty/Prefabs.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: b1fe1629894031f4fa5a650833c30cc6
-folderAsset: yes
-timeCreated: 1536607819
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
