--- Assets/Microgames/DrinkingParty/Scripts/DrinkingPartyScript.cs.meta
@@ -1,24 +1,0 @@
-fileFormatVersion: 2
-guid: 26557711dfa72b34a93fe1306a9d3255
-timeCreated: 1536611137
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences:
-  - ObjectLiquide1: {instanceID: 0}
-  - ObjectLiquide2: {instanceID: 0}
-  - ObjectCara1: {instanceID: 0}
-  - ObjectCara2: {instanceID: 0}
-  - Cara1_Anim1: {instanceID: 0}
-  - Cara1_Anim2: {instanceID: 0}
-  - Cara1_Anim3: {instanceID: 0}
-  - Cara2_Anim1: {instanceID: 0}
-  - Cara2_Anim2: {instanceID: 0}
-  - Cara2_Anim3: {instanceID: 0}
-  - DrinkSound: {instanceID: 0}
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
